# openapi.json
openapi: 3.0.3
info:
  title: Health Insurance Adviser API
  version: 1.0.0
paths:
  /api/v2/catalogue/:
    get:
      operationId: v2_catalogue_readiness
      tags:
      - v2
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CatalogueReadiness'
          description: ''
  /api/v2/comparisons/{id}/:
    get:
      operationId: v2_comparison_retrieve
      parameters:
      - in: path
        name: id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - v2
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ComparisonDetail'
          description: ''
  /api/v2/conversations/:
    get:
      operationId: v2_conversation_list
      tags:
      - v2
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConversationPage'
          description: ''
    post:
      operationId: v2_conversation_create
      tags:
      - v2
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/V2Conversation'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/V2Conversation'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/V2Conversation'
      security:
      - cookieAuth: []
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/V2Conversation'
          description: ''
  /api/v2/conversations/{conversation_id}/uploads/:
    post:
      operationId: v2_customer_upload
      parameters:
      - in: path
        name: conversation_id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - v2
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Upload'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/Upload'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Upload'
        required: true
      security:
      - cookieAuth: []
      responses:
        '202':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadAccepted'
          description: ''
  /api/v2/conversations/{id}/:
    get:
      operationId: v2_conversation_retrieve
      parameters:
      - in: path
        name: id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - v2
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/V2Conversation'
          description: ''
  /api/v2/conversations/{id}/messages/:
    get:
      operationId: v2_message_list
      parameters:
      - in: path
        name: id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - v2
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessagePage'
          description: ''
    post:
      operationId: v2_message_submit
      parameters:
      - in: path
        name: id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - v2
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MessageSubmission'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/MessageSubmission'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/MessageSubmission'
        required: true
      security:
      - cookieAuth: []
      responses:
        '202':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TurnAccepted'
          description: ''
  /api/v2/conversations/{id}/profile/:
    get:
      operationId: v2_profile_retrieve
      parameters:
      - in: path
        name: id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - v2
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CurrentProfile'
          description: ''
    patch:
      operationId: v2_profile_correct
      parameters:
      - in: path
        name: id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - v2
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchedProfileCorrection'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/PatchedProfileCorrection'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/PatchedProfileCorrection'
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CurrentProfile'
          description: ''
  /api/v2/documents/{id}/file/:
    get:
      operationId: v2_document_file_retrieve
      parameters:
      - in: path
        name: id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - v2
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/pdf:
              schema:
                type: string
                format: binary
          description: ''
  /api/v2/evidence/{id}/:
    get:
      operationId: v2_evidence_retrieve
      parameters:
      - in: path
        name: id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - v2
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EvidenceDetail'
          description: ''
  /api/v2/knowledge/readiness/:
    get:
      operationId: v2_knowledge_readiness
      tags:
      - v2
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CatalogueReadiness'
          description: ''
  /api/v2/turns/{id}/:
    get:
      operationId: v2_turn_retrieve
      parameters:
      - in: path
        name: id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - v2
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/V2Turn'
          description: ''
  /api/v2/turns/{id}/cancel/:
    post:
      operationId: v2_turn_cancel
      parameters:
      - in: path
        name: id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - v2
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/V2Turn'
          description: ''
  /api/v2/turns/{id}/events/:
    get:
      operationId: v2_turn_events
      parameters:
      - in: path
        name: id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - v2
      security:
      - cookieAuth: []
      responses:
        '200':
          description: Durable SSE events
  /api/v2/turns/{id}/retry/:
    post:
      operationId: v2_turn_retry
      parameters:
      - in: path
        name: id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - v2
      security:
      - cookieAuth: []
      responses:
        '202':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TurnAccepted'
          description: ''
  /api/v2/uploads/{id}/file/:
    get:
      operationId: v2_customer_upload_file_retrieve
      parameters:
      - in: path
        name: id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - v2
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/pdf:
              schema:
                type: string
                format: binary
          description: ''
...
components:
  schemas:
...
    CatalogueReadiness:
      type: object
      properties:
        catalogue_limit:
          type: integer
        comparison_label:
          type: string
        warning:
          type: string
        channel:
          type: string
        generation:
          type: integer
        release:
          allOf:
          - $ref: '#/components/schemas/KnowledgeRelease'
          nullable: true
        ready:
          type: boolean
        incomplete_comparison:
          type: boolean
        products:
          type: array
          items:
            $ref: '#/components/schemas/CatalogueProduct'
        interactive_routes:
          type: array
          items:
            $ref: '#/components/schemas/InteractiveRouteStatus'
        blocking_reason:
          type: string
          nullable: true
      required:
      - blocking_reason
      - catalogue_limit
      - channel
      - comparison_label
      - generation
      - incomplete_comparison
      - interactive_routes
      - products
      - ready
      - release
      - warning
...
    ComparisonDetail:
      type: object
      properties:
        id:
          type: string
          format: uuid
        turn_id:
          type: string
          format: uuid
        outcome:
          type: string
        profile_revision:
          type: integer
        knowledge_release_id:
          type: string
          format: uuid
        catalogue_limit:
          type: integer
        comparison_label:
          type: string
        created_at:
          type: string
          format: date-time
        products:
          type: array
          items:
            $ref: '#/components/schemas/ComparedProduct'
        information_needs:
          type: array
          items:
            $ref: '#/components/schemas/InformationNeed'
        statements:
          type: array
          items:
            $ref: '#/components/schemas/ComparisonStatement'
      required:
      - catalogue_limit
      - comparison_label
      - created_at
      - id
      - information_needs
      - knowledge_release_id
      - outcome
      - products
      - profile_revision
      - statements
      - turn_id
...
    ConversationPage:
      type: object
      properties:
        next:
          type: string
          format: uri
          nullable: true
        previous:
          type: string
          format: uri
          nullable: true
        results:
          type: array
          items:
            $ref: '#/components/schemas/V2Conversation'
      required:
      - next
      - previous
      - results
...
    MessagePage:
      type: object
      properties:
        next:
          type: string
          format: uri
          nullable: true
        previous:
          type: string
          format: uri
          nullable: true
        results:
          type: array
          items:
            $ref: '#/components/schemas/V2Message'
      required:
      - next
      - previous
      - results
    MessageSubmission:
      type: object
      properties:
        request_id:
          type: string
          format: uuid
        text:
          type: string
          maxLength: 20000
          minLength: 1
        expected_profile_revision:
          type: integer
          minimum: 1
      required:
      - request_id
      - text
...
    StatusEnum:
      enum:
      - open
      - archived
      - deleting
      - deleted
      type: string
      description: |-
        * `open` - open
        * `archived` - archived
        * `deleting` - deleting
        * `deleted` - deleted
    TurnAccepted:
      type: object
      properties:
        message_id:
          type: string
          format: uuid
        turn_id:
          type: string
          format: uuid
        event_url:
          type: string
        created:
          type: boolean
      required:
      - event_url
      - message_id
      - turn_id
...
    Upload:
      type: object
      properties:
        file:
          type: string
          format: uri
        kind:
          $ref: '#/components/schemas/KindEnum'
        source_message_id:
          type: string
          format: uuid
      required:
      - file
      - kind
    UploadAccepted:
      type: object
      properties:
        id:
          type: string
          format: uuid
        job_id:
          type: string
          format: uuid
        kind:
          type: string
        review_status:
          type: string
      required:
      - id
      - job_id
      - kind
      - review_status
    V2Conversation:
      type: object
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        title:
          type: string
          maxLength: 200
        status:
          allOf:
          - $ref: '#/components/schemas/StatusEnum'
          readOnly: true
        preferred_language:
          type: string
          nullable: true
          maxLength: 35
        profile_revision:
          type: integer
          readOnly: true
          nullable: true
        created_at:
          type: string
          format: date-time
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
      required:
      - created_at
      - id
      - profile_revision
      - status
      - updated_at
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: sessionid